# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: ScreenShare Backend API
  version: "1.0.0"
  description: >
    Contrato HTTP do backend do ScreenShare (issue #27). Cobre autenticação
    própria (cadastro, login, OAuth, refresh, logout), grupos, convites,
    presença e emissão de credenciais TURN. A sinalização WebRTC
    (offer/answer/ICE) usa WebSocket, definido separadamente na issue #38 —
    não faz parte deste documento. Ver docs/backend/ARQUITETURA.md pra
    decisões de implementação e status por issue.
servers:
  - url: /v1
    description: Versão atual. Mudanças incompatíveis usam /v2 em vez de alterar esta.
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Cadastrar com e-mail e senha
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 200 }
      responses:
        "201":
          description: Conta criada; sessão já emitida (login automático).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "409":
          description: E-mail já cadastrado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      tags: [auth]
      summary: Login com e-mail e senha
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Sessão emitida.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: E-mail ou senha inválidos — resposta genérica, não diferencia conta inexistente de senha errada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Trocar um refresh token válido por uma nova sessão
      operationId: authRefresh
      security: []
      description: O refresh token usado é revogado (rotação) — a resposta traz um novo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Nova sessão.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: Refresh token inválido, expirado ou já usado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/logout:
    post:
      tags: [auth]
      summary: Revogar um refresh token
      operationId: authLogout
      security: []
      description: Idempotente — token já revogado ou inexistente também responde 204.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: Refresh token revogado (ou já estava).
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/oauth/{provider}/start:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { type: string, enum: [google, github, discord] }
      - name: target
        in: query
        schema: { type: string, enum: [browser, desktop], default: browser }
        description: >
          Pra onde o callback deve redirecionar ao final: a página de
          fallback do navegador (dev) ou o deep link do app desktop
          empacotado. Ausente ou fora do allowlist cai em "browser". Viaja
          embutido no `state` assinado (não é preciso reenviar no
          callback).
    get:
      tags: [auth]
      summary: Inicia o login OAuth — redireciona pro provedor
      operationId: authOAuthStart
      security: []
      responses:
        "302":
          description: Redireciona pra URL de autorização do provedor, com um state assinado e expirável.
        "404":
          description: Provedor desconhecido ou sem credenciais configuradas ainda.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/oauth/{provider}/callback:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { type: string, enum: [google, github, discord] }
      - name: code
        in: query
        schema: { type: string }
      - name: state
        in: query
        schema: { type: string }
      - name: error
        in: query
        schema: { type: string }
        description: "Erro reportado pelo provedor (ex. usuário cancelou) — nunca repassado cru ao cliente."
    get:
      tags: [auth]
      summary: Callback do provedor OAuth — troca code por sessão própria
      operationId: authOAuthCallback
      security: []
      responses:
        "302":
          description: >
            Redireciona pro destino escolhido em `/start` via `?target=`
            (OAUTH_FRONTEND_REDIRECT_URL_BROWSER ou _DESKTOP; "browser" por
            padrão se o `state` não puder ser lido) com o resultado num
            fragmento da URL (nunca query string, pra não ficar em log):
            sucesso vira `#access_token=...&refresh_token=...&provider=...`;
            falha vira `#error=<código>` (`invalid_state`, `missing_code`,
            `provider_error`, `provider_not_configured`, `provider_unknown`).
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Access token de sessão emitido pelo próprio backend (issue #30, via
        /auth/register, /auth/login ou /auth/oauth/{provider}/callback) —
        não é mais um token do Firebase. Vida curta (15 min); renovar com
        /auth/refresh usando o refreshToken da mesma resposta. Payload
        (base64url, sem verificação de assinatura no cliente) traz
        `{uid, email?, displayName?, avatarUrl?, exp}` — os dois últimos
        (issue #70) só existem se a conta já fez login via OAuth alguma
        vez; login/refresh por e-mail+senha sem OAuth vinculado não tem
        nome/avatar do backend, o cliente decide o fallback (ex.: e-mail).
...
  responses:
...
    ValidationError:
      description: Corpo da requisição inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                error:
                  code: "validation_error"
                  message: "O campo 'name' é obrigatório."
                  correlationId: "b3f1c2d7"
    RateLimited:
      description: >
        Muitas requisições desse IP nesta rota — limite básico nas rotas
        públicas/sem sessão de /auth e em /turn-credentials (essa é
        autenticada, mas cada chamada gera uma requisição de verdade pra
        API da Cloudflare); não se aplica ao resto da API.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                error:
                  code: "rate_limited"
                  message: "Muitas requisições. Tente novamente em 1 minute."
                  correlationId: "b3f1c2d8"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              description: Identificador estável e machine-readable do erro.
            message:
              type: string
              description: Texto em pt-BR, seguro para exibir ao usuário.
            correlationId:
              type: string
              description: Correlaciona com os logs do backend (issue #44). Nunca contém dado sensível.
    Session:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: Bearer token de vida curta (15 min) — usar em toda requisição autenticada.
        refreshToken:
          type: string
          description: Valor opaco de vida longa (30 dias) — trocar por nova sessão em /auth/refresh antes de expirar. Rotaciona a cada uso.
...